Create utility function for deriving error messages
