+++ packages/peregrine/lib/talons/FormError/useFormError.js
@@ -1,19 +1,12 @@
 import { useMemo } from 'react';
+import { deriveErrorMessage } from '../../util/deriveErrorMessage';
 
 export const useFormError = props => {
     const { errors } = props;
 
-    const derivedErrorMessage = useMemo(() => {
-        const errorCollection = errors.filter(Boolean).map(error => {
-            const { graphQLErrors, message } = error;
-
-            return graphQLErrors && graphQLErrors.length
-                ? graphQLErrors.map(({ message }) => message).join(', ')
-                : message;
-        });
-
-        return errorCollection.join(', ');
-    }, [errors]);
+    const derivedErrorMessage = useMemo(() => deriveErrorMessage(errors), [
+        errors
+    ]);
 
     return {
         errorMessage: derivedErrorMessage
